01_development_environment_testing_framework/sdn-everything/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0
kubernetes>=27.2.0
docker>=6.0.0
requests>=2.31.0
bandit[toml]>=1.7.0
safety>=2.3.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
